requirements.txt
# Core LangChain stack
langchain==0.1.19
langchain-openai
langgraph==0.0.34
langsmith==0.1.21
openai==1.30.1

# Optional: Environment & tracing
python-dotenv==1.0.1

# Optional: Web UI (choose only one)
flask==3.0.3
# streamlit==1.35.0

# Logging and utility
rich==13.7.1

# For development/testing
ipykernel==6.29.4
